+++ analyzer/src/main/kotlin/com/github/cubuspl42/sigmaLang/analyzer/evaluation/values/SealedValue.kt
@@ -1,0 +1,3 @@
+package com.github.cubuspl42.sigmaLang.analyzer.evaluation.values
+
+abstract class SealedValue : Value
