+++ analyzer/src/main/kotlin/com/github/cubuspl42/sigmaLang/analyzer/semantics/types/IntCollectiveType.kt
@@ -1,0 +1,12 @@
+package com.github.cubuspl42.sigmaLang.analyzer.semantics.types
+
+object IntCollectiveType : IntType() {
+    override fun findLowestCommonSupertype(
+        other: Type,
+    ): Type = when (other) {
+        is IntType -> IntCollectiveType
+        else -> AnyType
+    }
+
+    override fun dump(): String = "Int"
+}
