+++ analyzer/src/main/kotlin/com/github/cubuspl42/sigmaLang/analyzer/semantics/types/IntType.kt
@@ -1,0 +1,3 @@
+package com.github.cubuspl42.sigmaLang.analyzer.semantics.types
+
+sealed class IntType : PrimitiveType()
