+++ analyzer/src/main/kotlin/com/github/cubuspl42/sigmaLang/analyzer/semantics/types/PrimitiveLiteralType.kt
@@ -1,0 +1,9 @@
+package com.github.cubuspl42.sigmaLang.analyzer.semantics.types
+
+import com.github.cubuspl42.sigmaLang.analyzer.evaluation.values.PrimitiveValue
+
+sealed interface PrimitiveLiteralType {
+    val asPrimitiveType: PrimitiveType
+
+    val value: PrimitiveValue
+}
