+++ gradle/libs.versions.toml
@@ -3,7 +3,7 @@
 annotations = "24.0.1"
 
 # plugins
-kotlin = "1.9.0"
+kotlin = "1.9.10"
 changelog = "2.1.2"
 gradleIntelliJPlugin = "1.15.0"
 qodana = "0.1.13"
